+++ src/core/libraries/np/np_web_api/np_web_api_internal.h
@@ -7,7 +7,7 @@
 #include "common/logging/log.h"
 #include "core/libraries/error_codes.h"
 #include "core/libraries/libs.h"
-#include "core/libraries/np/np_web_api.h"
+#include "np_web_api.h"
 
 namespace Libraries::Np::NpWebApi {
